✨ (install_nvim.sh): add user prompt for Neovim installation ♻️ (setup.sh): include install_nvim.sh in setup process

+++ setup/scripts/install_nvim.sh
@@ -1,20 +1,13 @@
-#!/bin/bash
+if ask "Do you want to install Neovim?"; then
+  rm -rf /opt/nvim
+  curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux-x86_64.appimage
+  chmod u+x nvim-linux-x86_64.appimage
+  mkdir -p /opt/nvim
+  mv nvim-linux-x86_64.appimage /opt/nvim/nvim
 
-BOLD='\033[1m'
-GREEN='\033[32m'
-YELLOW='\033[33m'
-RESET='\033[0m'
+  update-alternatives --install /usr/bin/vi vi /usr/bin/nvim 60
+  update-alternatives --install /usr/bin/vim vim /usr/bin/nvim 60
+  update-alternatives --install /usr/bin/editor editor /usr/bin/nvim 60
 
-echo -e "${BOLD}${YELLOW}Installing Neovim...${RESET}"
-
-rm -rf /opt/nvim
-curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux-x86_64.appimage
-chmod u+x nvim-linux-x86_64.appimage
-mkdir -p /opt/nvim
-mv nvim-linux-x86_64.appimage /opt/nvim/nvim
-
-update-alternatives --install /usr/bin/vi vi /usr/bin/nvim 60
-update-alternatives --install /usr/bin/vim vim /usr/bin/nvim 60
-update-alternatives --install /usr/bin/editor editor /usr/bin/nvim 60
-
-echo -e "${BOLD}${GREEN}Neovim installed.${RESET}"
+  echo -e "${BOLD}${GREEN}Neovim installed.${RESET}"
+fi
